Adjust logging format
