+++ src/renamarr/sonarr_py/adapter.py
@@ -22,8 +22,8 @@ from sonarr import CommandStatus as SonarrCommandStatus
 from sonarr.rest import ApiException
 
 from renamarr.exceptions import ArrOperationError
-from renamarr.models import (
-    CommandStatus,
+from renamarr.models.command import CommandStatus
+from renamarr.models.media import (
     FileRenameBatch,
     FileRenameCandidate,
     FolderRenameBatch,
