+++ tests/test_sonarr_py_adapter_contract.py
@@ -14,8 +14,8 @@ from sonarr import CommandStatus as SonarrCommandStatus
 from sonarr.rest import ApiException
 
 from renamarr.exceptions import ArrOperationError
-from renamarr.models import (
-    CommandStatus,
+from renamarr.models.command import CommandStatus
+from renamarr.models.media import (
     FileRenameBatch,
     FileRenameCandidate,
     FolderRenameBatch,
